Align logic across all binders.
The populators generateed by AnnotationBinder are usually the ones
executed at startup, but when --debug is passed, we fall back on
a generic populator. This logic did not have appropriate code to
update the method frame reads and writes in MethodIndex, and as a
result our --debug test runs (test:mri, spec:ruby) failed to set
up frames for methods that needed them.

This change re-aligns AnnotationBinder and the logic in RubyModule
as well as IndyBinder, but there's still a lot of duplication here
and potential for mistakes.