Revert to using weakref.rb, since it's based on a proper weak map.
This library used to use _id2ref (an internal API) for its
implementation. This was not only very VM-specific (most VMs
can't reconstitute an object reference given only its ID) but also
buggy (an evacuated reference might get overlaid by a new object
due to MRI's conservative GC and use of address for IDs.

The new implementation of the library uses ObjectSpace::WeakMap,
which is considered an internal API but which exposes an opaque
weak identity map that can be implemented on other VMs more
more easily. Given this and our need to continue descending from
Delegate, moving back to the Ruby version may make sense.