Fixes #4895. Regression in 9.1.15.0 with some ensure blocks being executed twice.
So I worked on a fairly big redo of how this all worked and it spiralled out
of control.  The solution I did eventually pick was nearly the simplest
possible one which would completely eliminate maintaining a stack of rescuepc
regions.

As it turns out maintaining the stack is largely flawed because any non-local
flow control can potentially leave a bunch of extra stack entries which are
no longer valid (they naturally only get popped from the stack if you encounter
an end region and non-local pretty much guarantees you won't).  If you then
hit a throw exception instr it will use that no longer valid stack info to
keep executing totally bogus stuff.

The solution is to basically replicate what we do when we make a CFG and process
start and end exception regions to generate a rescue ip map.  This means
walking all instrs an additional time to make the map but it eliminates the
stack and it makes figuring out rpc in the interpreter nearly a free operation
(remember we were constructing a stack and using it every call).

I tried to see if this extra rescue map generation led to slower startup and
I cannot tell.  It no doubt will help some infinite loop methods which never
make it into the JIT.