Made method storage backed by ByteList instead of String.  I knew this day would come but I was hoping I could cheat with 8859_1 strings enough to merge for 9.2.  Unfortunately, lack of encoding through method methods means we cannot accurately throw properly encoded error messages since he have long lost the encoding of the method.
So with that said, most people will not be able to notice this change.  All the
String methods for methods still exist.  These methods will assume all data is
UTF-8.  This could be a dicey gamble but since it is the default for Ruby it
likely will not impact anyone.  Also since UTF-8 is 7bit clean for ASCII we
should totally not have an issue there.

Some API signatures have changed.  This means 9.2 might end up causing some
issues if you are doing something very low-level with our code base (although
changes up to this point have already broken some APIs within parsing and IR
and that is unavoidable).  The method reader/writer maps needed their return
type changed from <String, DynamicMethod> to <ByteList, DynamicMethod>.  It
was an option to make a new name for methods which returned this but this is
so far down in our APIs I think we should run with it.  Also ProfiledMethod
has changed from String to ByteList (e.g. getName() -> ByteList from String).
Ultimately ProfiledMethod needs to be properly encoded to be able to report
the method in a properly encoded way...so there you have it.

Next phase is to change interpreters and the JIT to stop from going:
ByteList -> String -> ByteList.  I did not do that part here since this is
pretty large already.