local_variables needs to acquire properly encoded bytes to check their first and last character to see if it is a local variable.
In truth, this is pretty weird code.  IdUtils only works with String and without
doubling the code for ByteList there is a weakness since non Java Charset
strings will not work here.  IdUtils itself is also pretty weird since we really
just want variables which are not special and this method checks for all sorts
of things we know these strings can never be.  (FIXME: added)