Still a lot of missing changes for errors but more filled in.
const_defined? totally changed and only broken now due to missing proper logic
for printing out type names in error messages.  Outstanding is if the string
contains something the encoding thinks is unprintable it will display the type
in quotes and escape the unprintable character.  dumpString is nearly correct
but it will not correctly pick up \u0000 for \0 but instead print out \x00.

Landing for now since this is again a bigish delta and it advances things a bit
more.