Fix regressions after chr fixes.  uptoCommon was changing our single byte encodings because they were shared String buts setEncoding was changing their encoding.
This seems like a tiny fix but it has a large change in it.  If we setEncoding()
on a shared String it will actually stop sharing the underlying bytelist. In
practice, this could mean some extra bytelist creation behind the scenes.  What
it fixes is n strings sharing the same bytelist suddenly getting a new encoding
because one decided to have force_encoding called on it.

Making that tiny change also forced a couple of other changes because flags
for sharing are fragile due to order dependency when constructing Ruby Strings.

Also tweaked ordering of a conditional which wrankled my delicate sensibilities.