Begin propagating exception hierarchy down stack.
This forms the meat of wiring up the new parallel exception
hierarchy. All public constructors of RaiseException are now
deprecated and unused. RaiseException.from() forms take their
place and know how to properly construct a Ruby exception and
wrap it with an appropriate RaiseException subclass.

Three native exceptions are added, mapping to the Ruby exception
type of the same name: Exception, StandardError, and
SignalException.

One location in the code that caught RaiseException is now able
to catch StandardError (RubyNumeric coerce logic). There are
likely others.

See #4781.