Fixes #5066.  The & operator doesn't work with quoted symbols.
We were building a String with the entire bytelist buffer for symbols instead
of proper bounds.  We only saw this issue because :'to_s' goes through Ruby
String processing before being made into a Symbol.  Sort of a weird combo
but it is possible this bad code in Symbol.getName() could be causing other
issues?

[Note: this will only be effective fix for 9.1.  For 9.2 this is already fixed
because we have switched to a completely different way of handling encoded
identifiers.]