Make Block.type final.
This is cleanup to make simplifying block dispatch more
straightforward. The type field in Block was only set rarely, and
usually just once for a given block when setting it up as e.g.
lambda. All paths now just create a new block with the given type,
which opens up the possibility of Block subclasses specific to
each type (ProcBlock, LambdaBlock) and less conditional logic.

As a side effect of this cleanup, RubyProc fields are now all also
final and it no longer stores a duplicate Block.Type.