Restructure JIT to happen in one place each for methods and blocks
This introduces a third method and block top, the "OptInterpreter"
version. All methods and blocks in a normal run are MixedMode now,
aggregating both an Interpreted and a slot for an optimized or
jitted version. This commit also simplifies some call paths in
the IR block bodies and lowers several methods into
CompiledIRBlockBody to be more direct.