Attempt to use ByteList 2.0 with proper CharSequence logic.
The changes here move away from using CharSequence methods from
ByteList when the intent is to traverse bytes rather than decoded
characters. Places where a ByteList is not typed as such in code
(type is CharSequence or Object) or where it is concatenated into
a String using `+` were not detected by my search and may still
be incorrect.