Fixes #5084.
First attempt at using newest install4j.  This version we got an updated OSS
license for and it works with Java 9+.  After install4j updated our old file
nearly everything works except that media type changed to 'windows' from
'win32'.  This also changed the platform variable name so that had to be
tweaked.  I did a mock release and put ONLY java 9 on a windows machine and it
installs.  Huzzah