Restructure copy_stream to avoid temporary IO around IO-like.
Fixes #4796.

The logic here was wrapping any non-IO, non-String, IO-like object
with a readable or writable channel wrapper, itself wrapped in a
new IO object. The outer IO object was set to not autoclose, to
avoid closing the IO-like thing it contained. However these
semantics do not fit the fully-transient nature of the pseudo-
channel, and caused the ChannelFD created to wrap it to leak in
our FilenoUtil registry.

The new logic cleans up some dead paths, renames some variables
for clarity, and avoids using an IO wrapper when we simply need
a pseduo-channel. This avoids leaking ChannelFD and fixes #4796.