Final move to push RubySymbol a bit further through to IR.  This is a reasonable move since RubySymbol contains both the ByteList we need for proper display (for errors whatnot) and it also has the raw interned ID String we use for table access of variables and identifiers.
The second reasonable reason for this move is when we execute IR we have already
bootstrapped the runtime so it is safe to have symbols here.  By pushing them
this far forward we can also just return them for instrs where we need to
return a symbol.  Like in the case of a 'def foo;end'.

This only did put instrs.  Get next.