Fix #5138. jruby > 9.1.13.0 crashing on bundler when certain jvm flags are enabled.
Better title would be CFG merges and deletes a BB other BBs still think they
are pointing to.

This seems reasonable to me at the moment and I want to land so if someone is
enterprising they can try and find more problems.  I will have people look at
this.