Do not capture or propagate to main any Unrescuable exceptions.
Unrescuable exceptions in JRuby are used for internal operations,
usually involving the termination of a thread or of the running
program. They should not be captured as the terminal state of a
Ruby thread, nor should they be propagated across threads. This
was causing CI to terminate (with a success return code) on 9.1
and master.