Do not consider methods we will not be able to call.
This caused java.util.Properties.new.clone to fail because we try
to use the most general implementation of a given method in our
dispatch logic (which would be Object#clone) but did not omit
methods that can't be made accessible (Object#clone is protected
and not setAccessible on JDK9 without a warning). We would proceed
to try to call those inaccessible methods since they were the
most general impl, causing an exception.