Yay.  Ruby made some errors statically fail.  For next,break,redo:
  - will fail in top-level scope as SyntaxError instead of LJE
  - will fail in eval as SyntaxError (we largely did something weird)

In addition, yield will fail in toplevel scope with a SyntaxError.

This change is great but we still have an outstanding issue. We IRBuild method
bodies on demand.  So this will throw a syntaxerror much later than MRI. This
can be worked around but if we want to keep laziness we will need to mark
non-lazy methods which contain these keywords and compile those eagerly.  Also
we need to do that without rewalking the planet to look for those nodes.