First pass at passing proper class for super without impl search.
This mimicks how MRI maintains the current call's level in a given
class hierarchy.

Before this change, in order to support super from module methods
(a common decorator pattern) we would have to search the hierarchy
for the included module. This mechanism also failed to work
properly if the module appeared more than once in the hierarchy,
since the higher included module's super call would restart the
process above the lower included module.

In MRI, they return a second value from method searching: the
class (or included module wrapper) in the hierarchy whence the
method came. I mimic this by adding it to the CacheEntry we store
at a given class. This allows us to immediately super from the
correct level, as well as avoiding the infinite recursion issue
caused by our hierarchy search. In the future, it may be possible
to eliminate the "frame klazz" by just passing it through to any
super invocation rather than expecting it to be on the frame
stack.