Fixes #3688. Multibyte identifiers not marshaled correctly.
This has two fixes. Missed error messages not using str() method to properly
build up encoded error string.  Also we should ask for rubyName vs Name so
we get valid bytes.  There is a mystery I added a FIXME for where I don't
understand why we build up a RubyString by walking type containment then
we walk that string again looking for problems.  So there may be more work
here?  but if so we will open up a new issue for it.