Fixes #5162.  Thread::Backtrace::Location#base_label for blocks differs from MRI.
This fix is not quite what I would have liked but I think the other issue we
will need to consider at some point is that this data is not m17n safe.  So
when we decide to fix that we will need to mangle differently and probably will
be able to clean up this a bit.