Fixes #5163. When Thread::Backtrace::Location#base_label and #label are #== on MRI they are also #equal?
Our desire for laziness was the cause of this issue.  In the future we will
migrate to using symbol for methodNames to fully support M17n and this fix
ickiness will be able to be deleted.  Largely we use basename for all cases
except blocks.  In either case we save off the string we make so the equality
property is preserved.

With symbols the strings we have will always end up as the same string and
equality will also be true.  The amount of API changes is a bit too much to
do the m17n work now.