All is not well in full interp/JIT land for block dispatch of lambdas.
ACP pass calls PrepareFixedBlockArgsInstr but each_with_index and a couple
of others when passed a lambda end up expecting the ary spreading we used to
do in prepareFixedBlockArgs.  I tried figuring out how these native calls
could be changed but I see both call and yield on block being hit.  Changing
the error in call to yieldValues did fix one case but I have no confidence
this is correct.

So I removed preparefixed for more generic prepareblockargs for now.  We need
to dig in more here but block args are so complicated.