Revert CharSequence compatibility change to #length.
This came from jruby/bytelist@190c465e9ecc681ec59699648fd3fcd761abd105
and was a first step in making ByteList conform to CharSequence,
for potentially cheaper Java integration.

The change to #length was too drastic to introduce along with the
reabsorption of ByteList, since third-party extensions may be
using it under its "byte length" definition.

We may be able to introduce this change in the future (9.3?) but
would need to work with all extension authors to get backward-
compatible patched versions released. It is possible, but not
likely to happen in the short term.

The Java integration gains might be partially achievable with a
minimal CharSequence wrapper *similar* to ByteList so that only
copying, rather than transcoding, would be needed to make the
call.