This code appears to be dead. Remove it to resolve #315.
Note that the arguments passed to the IRRuntimeHelpers method
restoreExceptionVar were both IRubyObject, and the only work done
in the method would be if the first one were an instance of
IRReturnJump and IRBreakJump. However neither of these classes
implements IRubyObject, and they are not extended. Similar logic
was removed in the interpreter. Local test:jruby and
spec:ruby:fast passed muster.