Use JDK atomic move to more closely emulate POSIX rename.
Fixes #5318.

Note that this no longer does the explicit file store check,
because that failed in an unrelated way when running under a
FreeBSD jail. Instead this now relies on the Files subsystem of
JDK to percolate out the original error message, which we should
detect properly and turn into EXDEV.