Make EQQInstr be a CallInstr.  This will be used in part to make JIT more opaque as this is the only consumer which does not extend CallBase for compileCallCommon.
EQQInstr is either an actual call or we fast path but from IR analysis
perspective it is reasonable for it to extend a call since it can perform one.

The rational for this change is that the JIT breaks all the elements out
of eqqinstr and callbase to compile calls.  I need hasliteralclosure where
the JIT lazily constructs callsites so by making all things which invoke
a callbase then we can pass that down there without having to change n
potential signatures within the JIT.