Use --module-path for JRuby when running on a jmods JDK.
By moving JRuby into the module path, we are able to silence
some of the reflection warnings. This will also be necessary
for users that want to expose core JDK classes to JRuby for
compatibility purposes.

The logic here looks in $JAVA_HOME for the 'jmods' dir, which
should be there for all OpenJDK 9+ and not in earlier releases.
If that dir is present, we alter the command line in the follwing
ways:

* If running verified, a la VERIFY_JRUBY=1, all jars including
  JRuby will be loaded via --module-path instead of -classpath.
* If running unverified, a la bootclasspath, the bootclasspath
  flag is replaced by --module-path, and an additional flag
  -verify:none is added to  blunt the cost of booting. This is not
  ideal, but all libraries we ship with and all bytecode we
  generate have been verified many times over in testing.

Note that these only affect launching JRuby from the dist release.
We will need similar changes to the .cmd launcher and the native
jruby-launcher.