Narrow change to fix for glob logic for #5333 to only escaped segs
This narrows my fix to only fire for segments with magic chars
(* and ? and the like) or backslashes in the presence of escaped
mode (i.e. not FNM_NOESCAPE). This will slow down the processing
of segments that contain only backslash escapes, but they were not
processed properly before anyway. Segments without backslashes or
magic characters are unaffected. Segments with backslashes and
magic characters are unaffected.