Attempt to fix windows oddity where some times a rename during bundle install would fail.
The fix is I think if we depend on Files.move to do it's thing we should not
try and rename a second time using the old way.  This happened to never
manifest on linux or macos because bundler stores files in another FS so
we never hit this code path.