Remove encoding from ByteList hash calc here.
We would like to have the table consider encoding, but until we
only use ByteList for identifiers this is not possible. We still
use Java's String in many places, in which case we only have
access to the raw bytes and not the encoding they should be
associated with, so the String-based lookup would never be able
to add this additional mask.

Unfortunately removing this additional hash dimension breaks one
Symbol marshal spec, but it was broken before this work as well.

This was the cause of Kernel#eval regressions in #4576.