Remove explicit Site checks to propagate callSiteId and moves that into the instrs clone().  It was confusing whether all paths would end up cloning and maintaining same callsiteid.
Realized that callsite itself was remade over and over and that would increment
callsiteid and make a new callsite.  Cloned version of the same call in a
different version of the same scope is the same call site.  Even if two versions
of the same scope were executing on two threads their type info regarding that
site is still valid from either form.

Beyond that justification the bug fixed was a scope in -X-C was making a new
profiled caching callsite but was retaining the temporary old callsiteid which
was being changed later by Site.setCallSiteId which was then not propagated into
its new callsite.