Taint ARGV values.
This is a new test for ARGV tainting that we never would have
passed, because we don't taint ARGV strings.

Note that MRI always taints all strings that come from "external"
bytes, which is similar to our "external" Java strings. We are
opting to localize this fix to ARGV only due to the unknown impact
of suddenly tainting all strings imported from Java via our
version of the "external" string construction.

https://bugs.ruby-lang.org/issues/14941