Limit the width of reified objects to 50 by default.
This relates to #5511 but is not a fix. This only limits how wide
these object will be, so we're not generating extremely wide
classes for weirdly dynamic objects that create lots of new
instance variables.

This will be improved once we make objects know about their own
shapes, so they can evolve to handle the right number of variables
that are actually used.