Fixes #5517.
An extra thing corrected here was that most returns within END blocks can be
statically determined so an LJE replaces the return during build time.  This
was a minor part of this but I mention it to get it out of the way.

The main fix was to change how we initiate a non-local return from a closure.
This commit changes from primarily walking a dynamic scope stack to walking
the static scope stack.   There are two reasons for this:

1. visibility to scopes which we do not add variables for
2. JITing may eliminate dynamic scopes

So the new mechanism uses getContainingReturnToScope.  As an added bonus
both checkForLJE and initiateNonLocalReturn use the same method now.  Before
the two methods were very very similar but one used IR and the other used
dynamic scope.

There is a tiny wrinkle to the new way but it does make sense.  static scope
is static but not all information we need is static.  We cannot know if a
closure will be a proc or a lambda.  So when we initiate a non-local return
so long as we are within closures we check dynamic scopes looking to see if
we are within a lambda (getContainingLambda).

My only wonderment of this whole commit is runInterpreter added a catch to
for IRReturnJump.  This is definitely ok to have here but I wonder if it
covers all paths into execution or not?  Is there another place where we should
add a catch?